Adding some relationships to the DB using Entity Framework

+++ Services/CharacterService/CharService.cs
@@ -1,8 +1,10 @@
 using System;
 using System.Collections.Generic;
 using System.Linq;
+using System.Security.Claims;
 using System.Threading.Tasks;
 using AutoMapper;
+using Microsoft.AspNetCore.Http;
 using Microsoft.EntityFrameworkCore;
 using try5000rpg.Data;
 using try5000rpg.DTOs.Characters;
@@ -16,20 +18,26 @@ namespace try5000rpg.Services.CharacterService
 
         private readonly IMapper _mapper;
         private readonly DataContext _context;
-
-        public CharService(IMapper mapper, DataContext context)
+        private readonly IHttpContextAccessor _HttpContextAccessor;
+        public CharService(IMapper mapper, DataContext context, IHttpContextAccessor httpContextAccessor)
         {
             _mapper = mapper;
             _context = context;
+            _HttpContextAccessor = httpContextAccessor;
         }
+
+        private int GetUserID() => int.Parse(_HttpContextAccessor.HttpContext.User.FindFirstValue(ClaimTypes.NameIdentifier));
 
         public async Task<ServiceResponse<List<GetCharacterDTO>>> CreateSingle(AddCharacterDTO newSingle)
         {
             var srListService = new ServiceResponse<List<GetCharacterDTO>>();
             Characters NewCharacter = _mapper.Map<Characters>(newSingle);
+            NewCharacter.User = await _context.Users.FirstOrDefaultAsync(u => u.Id == GetUserID());
             _context.Characters.Add(NewCharacter);
             await _context.SaveChangesAsync();
-            srListService.Data = await _context.Characters.Select(c => _mapper.Map<GetCharacterDTO>(c)).ToListAsync();
+            srListService.Data = await _context.Characters
+                .Where(c => c.User.Id == GetUserID())
+                .Select(c => _mapper.Map<GetCharacterDTO>(c)).ToListAsync();
             return srListService;
         }
 
@@ -38,10 +46,20 @@ namespace try5000rpg.Services.CharacterService
             var srResponse = new ServiceResponse<List<GetCharacterDTO>>();
             try
             {
-                Characters chSingleCharacter = await _context.Characters.FirstAsync(c => c.Id == iCharacterID);
-                _context.Characters.Remove(chSingleCharacter);
-                await _context.SaveChangesAsync();
-                srResponse.Data = _context.Characters.Select(c => _mapper.Map<GetCharacterDTO>(c)).ToList();
+                Characters chSingleCharacter = await _context.Characters.FirstOrDefaultAsync(c => c.Id == iCharacterID && c.User.Id == GetUserID());
+                if (chSingleCharacter != null)
+                {
+                    _context.Characters.Remove(chSingleCharacter);
+                    await _context.SaveChangesAsync();
+                    srResponse.Data = _context.Characters
+                        .Where(c => c.User.Id == GetUserID())
+                        .Select(c => _mapper.Map<GetCharacterDTO>(c)).ToList();
+                }
+                else
+                {
+                    srResponse.Success = false;
+                    srResponse.Message = "No Character";
+                }
                 
             }
             catch (Exception ex)
@@ -52,10 +70,11 @@ namespace try5000rpg.Services.CharacterService
             return srResponse;
         }
 
-        public async Task<ServiceResponse<List<GetCharacterDTO>>> GetAllCharacters(int UserId)
+        public async Task<ServiceResponse<List<GetCharacterDTO>>> GetAllCharacters()
         {
+            
             var srListService = new ServiceResponse<List<GetCharacterDTO>>();
-            var dbCharacters = await _context.Characters.Where(c => c.User.Id == UserId).ToListAsync();
+            var dbCharacters = await _context.Characters.Where(c => c.User.Id == GetUserID()).ToListAsync();
             srListService.Data = dbCharacters.Select(c => _mapper.Map<GetCharacterDTO>(c)).ToList();
             return srListService;
         }
@@ -63,7 +82,7 @@ namespace try5000rpg.Services.CharacterService
         public async Task<ServiceResponse<GetCharacterDTO>> GetSingle(int id)
         {
             var srSingleCharacter = new ServiceResponse<GetCharacterDTO>();
-            var drSingleCharacter = await _context.Characters.FirstOrDefaultAsync(c => c.Id == id);
+            var drSingleCharacter = await _context.Characters.FirstOrDefaultAsync(c => c.Id == id && c.User.Id == GetUserID());
             srSingleCharacter.Data = _mapper.Map<GetCharacterDTO>(drSingleCharacter);
             return srSingleCharacter;
         }
@@ -73,14 +92,24 @@ namespace try5000rpg.Services.CharacterService
             var srResponse = new ServiceResponse<GetCharacterDTO>();
             try
             {
-                Characters chSingleCharacter = await _context.Characters.FirstOrDefaultAsync(c => c.Id == updatedCharacter.Id);
-                chSingleCharacter.Name = updatedCharacter.Name;
-                chSingleCharacter.Class = updatedCharacter.Class;
-                chSingleCharacter.Defense = updatedCharacter.Defense;
-                chSingleCharacter.Intelligance = updatedCharacter.Intelligance;
-                chSingleCharacter.Strength = updatedCharacter.Strength;
-                await _context.SaveChangesAsync();
-                srResponse.Data = _mapper.Map<GetCharacterDTO>(chSingleCharacter);
+                Characters chSingleCharacter = await _context.Characters
+                    .Include(c => c.User)
+                    .FirstOrDefaultAsync(c => c.Id == updatedCharacter.Id);
+                if (chSingleCharacter.User.Id == GetUserID())
+                {
+                    chSingleCharacter.Name = updatedCharacter.Name;
+                    chSingleCharacter.Class = updatedCharacter.Class;
+                    chSingleCharacter.Defense = updatedCharacter.Defense;
+                    chSingleCharacter.Intelligance = updatedCharacter.Intelligance;
+                    chSingleCharacter.Strength = updatedCharacter.Strength;
+                    await _context.SaveChangesAsync();
+                    srResponse.Data = _mapper.Map<GetCharacterDTO>(chSingleCharacter);
+                }
+                else
+                {
+                    srResponse.Success = false;
+                    srResponse.Message = "no character";
+                }
             }
             catch (Exception ex)
             {
